+++ NEWS.md
@@ -8,12 +8,12 @@
   
 # inti 0.6.4
 
+- Update bootstrap in apps
 - Tarpuy:
   - Alows to exclude or delete `{when}` or `{sample}` colums
-  - Allow `defaultValue` for numeric traits
+  - Allow `defaultValue` for traits
 - Rticles
   - Include images using markdown syntax `![]()`
-  
   
 => Fix if "defaultvalue" is in the trait table